# diff element not optional