--- a/reactpy_vdom/vdom.mojo
+++ b/reactpy_vdom/vdom.mojo
@@ -19,21 +19,19 @@ struct Element(Copyable, Movable, Defaultable, EqualityComparable, Stringable, R
         self.text = Optional[String](None)
         self.key = Optional[String](None)
 
-    fn __init__(out self, tag: String, attributes: Dict[String, String], children: List[Element], key: Optional[String] = None):
+    fn __init__(out self, owned tag: String, owned attributes: Dict[String, String], owned children: List[Element], owned key: Optional[String] = None):
         self.tag = tag
         self.attributes = attributes
         self.children = children
         self.text = Optional[String](None)
         self.key = key
-        # self.key = Optional[String](None)
 
-    fn __init__(out self, tag: String, attributes: Dict[String, String], text: String, key: Optional[String] = None):
+    fn __init__(out self, owned tag: String, owned attributes: Dict[String, String], owned text: String, owned key: Optional[String] = None):
         self.tag = tag
         self.attributes = attributes
         self.children = []
         self.text = Optional(text)
         self.key = key
-        # self.key = Optional[String](None)
 
     fn __eq__(self, other: Self) -> Bool:
         return self.tag == other.tag and dict_equal(self.attributes, other.attributes) and self.children == other.children and self.text == other.text and self.key == other.key
@@ -64,21 +62,6 @@ struct Element(Copyable, Movable, Defaultable, EqualityComparable, Stringable, R
         s += "</" + self.tag + ">"
 
         return s
-
-    # fn get(ref self, path: List[Int]) raises -> ref [self] Self:
-    #     """Get node at path."""
-    #     var current = Pointer(to=self)
-    #     for index in path:
-    #         debug_assert(-len(current[].children) <= index < len(current[].children), "index: ",
-    #             index,
-    #             " is out of bounds for `children` of length: ",
-    #             len(current[].children))
-    #         if index < 0 or index >= len(current[].children):
-    #             current_ = Pointer(to=current[].children[index])
-    #             current = rebind[Pointer[Element, __origin_of(self)]](current_)
-    #         else:
-    #             raise Error("index: ", index, " is out of bounds for `children` of length: ", len(current[].children))
-    #     return current[]
 
     fn get(ref self, path: List[Int]) raises -> ref [self] Self:
         """Get node at path."""